third_party/blink/web_tests/fast/events/before-unload-returnValue-expected.txt

CONFIRM NAVIGATION
Tests the returnValue attribute of the BeforeUnloadEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testEvent.__proto__ is BeforeUnloadEvent.prototype
PASS testEvent.returnValue is ""
PASS testEvent.returnValue is "This is beforeunload from the top level frame."
PASS successfullyParsed is true

TEST COMPLETE
